+++ CHANGELOG.md
@@ -5,6 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.1] - 2025-10-21
+
+### Added
+
+- Memorize components to not re-render if not property changes.
+- Rollup-plugin-visualizer to analize bundle library size.
+
+### Removed
+
+- Removed Luxon library and types.
+
+### Changed
+
+- Use native Date JS class to get current year.
+
+### Fixed
+
+-
+
 ## [1.1.0] - 2025-10-20
 
 ### Changed